# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Autonomous Job Search Agent",
    "description": "Ingest public Greenhouse and Lever listings, run deterministic hard filters, evaluate jobs with JobEvaluationAgent, and review results in the dashboard.\n\nThe review UI is served at `/`. Interactive docs are at `/docs` (Swagger UI) and `/redoc`. Human Approve/Review/Reject decisions do not submit applications.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/jobs": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Create job",
        "description": "Insert a listing. Duplicate source + source_job_id returns 409.",
        "operationId": "post_job_jobs_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List jobs",
        "description": "Filter and paginate ingested listings.",
        "operationId": "get_jobs_jobs_get",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobSource"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "name": "title",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Title"
            }
          },
          {
            "name": "location",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location"
            }
          },
          {
            "name": "remote_policy",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RemotePolicy"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Remote Policy"
            }
          },
          {
            "name": "seniority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Seniority"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Seniority"
            }
          },
          {
            "name": "minimum_salary",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Minimum Salary"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get job",
        "description": "Return a single listing.",
        "operationId": "get_job_by_id_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/review/jobs": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "Review job queue",
        "description": "Evaluated jobs for human review. Filter by recommendation, score, source, remote, date.",
        "operationId": "get_review_queue_review_jobs_get",
        "parameters": [
          {
            "name": "recommendation",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Recommendation"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Recommendation"
            }
          },
          {
            "name": "minimum_score",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Minimum Score"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobSource"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "name": "remote_policy",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RemotePolicy"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Remote Policy"
            }
          },
          {
            "name": "discovered",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Discovered"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobQueueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/review/jobs/{job_id}": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "Review job detail",
        "description": "Description, evaluation, cited evidence, hard-filter result, and application URL.",
        "operationId": "get_review_job_review_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobReviewDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/review/jobs/{job_id}/decision": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Record a human decision",
        "description": "Store approve, review, or reject. Does not submit an application.",
        "operationId": "post_review_decision_review_jobs__job_id__decision_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HumanDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobReviewDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sources/greenhouse/{board_token}/sync": {
      "post": {
        "tags": [
          "sources"
        ],
        "summary": "Sync Greenhouse board",
        "description": "Fetch `boards-api.greenhouse.io` jobs for a public board token and upsert.",
        "operationId": "sync_greenhouse_board_sources_greenhouse__board_token__sync_post",
        "parameters": [
          {
            "name": "board_token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "pattern": "^[A-Za-z0-9_-]+$",
              "description": "Public job-board slug",
              "title": "Board Token"
            },
            "description": "Public job-board slug"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceSyncResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sources/lever/{site}/sync": {
      "post": {
        "tags": [
          "sources"
        ],
        "summary": "Sync Lever site",
        "description": "Fetch `api.lever.co` postings for a public site slug and upsert.",
        "operationId": "sync_lever_site_sources_lever__site__sync_post",
        "parameters": [
          {
            "name": "site",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "pattern": "^[A-Za-z0-9_-]+$",
              "description": "Public job-board slug",
              "title": "Site"
            },
            "description": "Public job-board slug"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceSyncResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "EmploymentType": {
        "type": "string",
        "enum": [
          "full_time",
          "contract",
          "part_time"
        ],
        "title": "EmploymentType"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobCreate": {
        "properties": {
          "source": {
            "$ref": "#/components/schemas/JobSource",
            "default": "manual"
          },
          "source_job_id": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Source Job Id"
          },
          "company": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Company"
          },
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "minLength": 1,
            "title": "Description"
          },
          "location": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "remote_policy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RemotePolicy"
              },
              {
                "type": "null"
              }
            ]
          },
          "employment_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EmploymentType"
              },
              {
                "type": "null"
              }
            ]
          },
          "seniority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Seniority"
              },
              {
                "type": "null"
              }
            ]
          },
          "salary_min": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Min"
          },
          "salary_max": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Max"
          },
          "salary_currency": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 3,
                "minLength": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Currency"
          },
          "job_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Url"
          },
          "application_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2048
              },
              {
                "type": "null"
              }
            ],
            "title": "Application Url"
          },
          "department": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Department"
          },
          "posted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posted At"
          },
          "discovered_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discovered At"
          },
          "raw_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Raw Data"
          }
        },
        "type": "object",
        "required": [
          "source_job_id",
          "company",
          "title",
          "description"
        ],
        "title": "JobCreate"
      },
...
      "JobListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "JobListResponse"
      },
...
      "JobQueueResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobQueueItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "JobQueueResponse"
      },
      "JobRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "source": {
            "$ref": "#/components/schemas/JobSource"
          },
          "source_job_id": {
            "type": "string",
            "title": "Source Job Id"
          },
          "company": {
            "type": "string",
            "title": "Company"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "remote_policy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RemotePolicy"
              },
              {
                "type": "null"
              }
            ]
          },
          "employment_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EmploymentType"
              },
              {
                "type": "null"
              }
            ]
          },
          "seniority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Seniority"
              },
              {
                "type": "null"
              }
            ]
          },
          "salary_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Min"
          },
          "salary_max": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Max"
          },
          "salary_currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Salary Currency"
          },
          "job_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Url"
          },
          "application_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Application Url"
          },
          "department": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Department"
          },
          "posted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Posted At"
          },
          "discovered_at": {
            "type": "string",
            "format": "date-time",
            "title": "Discovered At"
          },
          "raw_data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Raw Data"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "is_duplicate_description": {
            "type": "boolean",
            "title": "Is Duplicate Description"
          },
          "duplicate_description_job_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Duplicate Description Job Ids"
          },
          "human_decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HumanDecision"
              },
              {
                "type": "null"
              }
            ]
          },
          "human_decided_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Human Decided At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "source_job_id",
          "company",
          "title",
          "description",
          "location",
          "remote_policy",
          "employment_type",
          "seniority",
          "salary_min",
          "salary_max",
          "salary_currency",
          "job_url",
          "application_url",
          "department",
          "posted_at",
          "discovered_at",
          "raw_data",
          "content_hash",
          "is_duplicate_description",
          "duplicate_description_job_ids",
          "created_at",
          "updated_at"
        ],
        "title": "JobRead"
      },
      "JobReviewDetail": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/JobRead"
          },
          "evaluation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobEvaluationRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "supporting_evidence": {
            "items": {
              "$ref": "#/components/schemas/CandidateEvidenceRead"
            },
            "type": "array",
            "title": "Supporting Evidence"
          },
          "hard_filter": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HardFilterResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "application_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Application Url"
          },
          "human_decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/HumanDecision"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "job",
          "evaluation",
          "supporting_evidence",
          "hard_filter",
          "application_url",
          "human_decision"
        ],
        "title": "JobReviewDetail"
      },
      "JobSource": {
        "type": "string",
        "enum": [
          "manual",
          "greenhouse",
          "lever",
          "ashby",
          "linkedin",
          "indeed",
          "company_site",
          "other"
        ],
        "title": "JobSource",
        "description": "Origin of an ingested listing. External adapters are not implemented yet."
      },
...
      "Recommendation": {
        "type": "string",
        "enum": [
          "apply",
          "review",
          "reject"
        ],
        "title": "Recommendation"
      },
      "RemotePolicy": {
        "type": "string",
        "enum": [
          "remote",
          "hybrid",
          "onsite",
          "unknown"
        ],
        "title": "RemotePolicy"
      },
...
      "Seniority": {
        "type": "string",
        "enum": [
          "junior",
          "mid",
          "senior",
          "staff",
          "principal",
          "distinguished"
        ],
        "title": "Seniority"
      },
...
    }
  },
...
}